@@ -38,11 +38,11 @@ task clean(type: Delete) {
 }
 
 ext {
-    compileSdkVersion = 28
-    buildToolsVersion = '25.0.0'
-    targetSdkVersion = 28
+    compileSdkVersion = 29
+    buildToolsVersion = '29.0.0'
+    targetSdkVersion = 29
     minSdkVersion = 23
     versionCode = 1
     versionName = "1.0"
-    supportVersion = '27.1.1'
+    supportVersion = '29.0.0'
 }
