@@ -1,9 +1,10 @@
 package com.yc.mugua.weight;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
+
+import androidx.viewpager.widget.ViewPager;
 
 public class NoScrollViewPager extends ViewPager {
     private boolean noScroll = false;
